fix: better error handling inheritance bug
