--- src/Examples/JsonApiDotNetCoreExample/Controllers/KebabCasedModelsController.cs
@@ -1,18 +1,0 @@
-using JsonApiDotNetCore.Configuration;
-using JsonApiDotNetCore.Controllers;
-using JsonApiDotNetCore.Services;
-using JsonApiDotNetCoreExample.Models;
-using Microsoft.Extensions.Logging;
-
-namespace JsonApiDotNetCoreExample.Controllers
-{
-    public sealed class KebabCasedModelsController : JsonApiController<KebabCasedModel>
-    {
-        public KebabCasedModelsController(
-            IJsonApiOptions options,
-            ILoggerFactory loggerFactory,
-            IResourceService<KebabCasedModel> resourceService)
-            : base(options, loggerFactory, resourceService)
-        { }
-    }
-}
